qr code
This project is a simple QR Code Generator built using Python and Tkinter for creating QR codes from user input (text or URL). The application provides an intuitive graphical user interface (GUI) to generate and save QR codes locally as images. It combines Tkinter for the interface and the qr code library for generating the QR codes.